+++ src/utils/security/reporting/index.ts
@@ -1,0 +1,7 @@
+/**
+ * セキュリティレポーティングシステム統合エクスポート
+ * 後方互換性のため、すべてのクラスと関数を再エクスポート
+ */
+export { SecurityReportGenerator } from './generator'
+export { SecurityReportDistributor } from './distributor'
+export { initializeSecurityReporting } from './initializer'
