+++ src/utils/security/reporting/initializer.ts
@@ -1,0 +1,11 @@
+import { SecurityReportDistributor } from './distributor'
+
+/**
+ * セキュリティレポーティングシステムの初期化
+ */
+export function initializeSecurityReporting(): void {
+  const distributor = SecurityReportDistributor.getInstance()
+  distributor.startScheduledReports()
+
+  // logger.debug('📊 Security reporting system initialized')
+}
